+++ open-sphere-base/core/src/main/java/io/opensphere/core/common/math/LUDecomposition.java
@@ -267,7 +267,7 @@ public class LUDecomposition implements java.io.Serializable
         double[] vals = new double[m];
         for (int i = 0; i < m; i++)
         {
-            vals[i] = (double)piv[i];
+            vals[i] = piv[i];
         }
         return vals;
     }
@@ -285,7 +285,7 @@ public class LUDecomposition implements java.io.Serializable
         {
             throw new IllegalArgumentException("Matrix must be square.");
         }
-        double d = (double)pivsign;
+        double d = pivsign;
         for (int j = 0; j < n; j++)
         {
             d *= LU[j][j];
